Tested the implementation of the k-NN algorithms
The k-NN algorithms should theoretically run. We need to talk about
what we're supposed to do with the fact that the labels on the data
are discrete unlike the k-NN data set in the homework.

We also need to talk about what we should do about the data being
so large. It's hard to do anything.